chore: refactor tests (#59)
