+++ examples/publish subscribe/GenerateAndBuild.sh
@@ -1,3 +1,4 @@
 #!/bin/sh
 
+rm -rf ./streetlight
 ../../node_modules/@asyncapi/generator/cli.js --output "./streetlight" "./streetlight.json" "../../" --force-write --param "generateTestClient=true" --param "promisifyReplyCallback=true" && cd ./streetlight && npm i && npm run build && cd ..
