+++ template/src/testclient/testchannels/$$channel$$.ts.js
@@ -1,10 +1,10 @@
 import { File } from '@asyncapi/generator-react-sdk';
-import { Publish } from '../../../../../components/channel/publish';
-import { Subscribe } from '../../../../../components/channel/subscribe';
-import { Reply } from '../../../../../components/channel/reply';
-import { Request } from '../../../../../components/channel/request';
-import { General } from '../../../../../components/channel/general';
-import { pascalCase, isRequestReply, isReplier, isRequester, isPubsub} from '../../../../../utils/index';
+import { Publish } from '../../../../components/channel/publish';
+import { Subscribe } from '../../../../components/channel/subscribe';
+import { Reply } from '../../../../components/channel/reply';
+import { Request } from '../../../../components/channel/request';
+import { General } from '../../../../components/channel/general';
+import { pascalCase, isRequestReply, isReplier, isRequester, isPubsub} from '../../../../utils/index';
 
 /**
  * Return the correct channel component based on whether its `pubSub` or `requestReply`.
@@ -65,7 +65,7 @@ export default function channelRender({ asyncapi, channelName, channel, params }
 
   return <File name={`${pascalCase(channelName)}.ts`}>
     {`
-${General(channel, publishMessage, subscribeMessage, '../../..')}
+${General(channel, publishMessage, subscribeMessage, '../..')}
 
 ${getChannelCode(asyncapi, channel, channelName, params)}
 `}
